Implemented UI w/ buttons
Created a canvas with background and icons overlayed, with clickable buttons on each icon. This is the groundwork for being able to pull up windows for the user to be able to access different parts of our game. Still working out the best way to pull up/hide these windows.